Move AnalyticsService to it's original place
This is for review purposes and should be reverted later.

COAND-844
